details page now displaying correctly and book form working correctly
