+++ frontend/src/Componets/BookCard.tsx
@@ -3,7 +3,7 @@ import {Book} from "../Book.ts";
 export default function BookCard(book:Book){
     return(
         <>
-        <h2>{book.titel}</h2>
+        <h2>{book.title}</h2>
          <img src={book.image} />
          <p>author: {book.author}</p>
         <p>isbn:{book.isbn}</p>
